+++ utils/credentials/credentials.go
@@ -10,14 +10,14 @@ import (
 	"github.com/ethereum/go-ethereum/crypto"
 )
 
-func TransactOptsFromPrivateKey(pk *ecdsa.PrivateKey, chainID int) (*bind.TransactOpts, error) {
+func TransactOptsFromPrivateKey(pk *ecdsa.PrivateKey, chainID int64) (*bind.TransactOpts, error) {
 	opts, _, err := CredentialsFromPrivateKey(pk, chainID)
 	return opts, err
 }
 
-func CredentialsFromPrivateKey(pk *ecdsa.PrivateKey, chainID int) (*bind.TransactOpts, *ecdsa.PrivateKey, error) {
+func CredentialsFromPrivateKey(pk *ecdsa.PrivateKey, chainID int64) (*bind.TransactOpts, *ecdsa.PrivateKey, error) {
 	opts, err := bind.NewKeyedTransactorWithChainID(
-		pk, big.NewInt(int64(chainID)),
+		pk, big.NewInt(chainID),
 	)
 	if err != nil {
 		return nil, nil, fmt.Errorf("bind.NewKeyedTransactorWithChainID: %w", err)
